requirements.txt
djangorestframework==3.10.3
requests
httpie==1.0.3
